Rename controllers package and Greeting record
Rename the controllers -> controller (singular), and Greeting -> JsonGreeting to
indicate that it is an object for use in the web layer
