+++ src/main/java/com/wilkins/showcase/controller/GreetingController.java
@@ -1,4 +1,4 @@
-package com.wilkins.showcase.controllers;
+package com.wilkins.showcase.controller;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -8,18 +8,18 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/greeting")
+@RequestMapping("/greetings")
 public class GreetingController {
 
     private static final Logger log = LoggerFactory.getLogger(GreetingController.class);
 
     @GetMapping
-    public Greeting getGreeting(@RequestParam(name = "salutation", required = false, defaultValue = "hello") String salutationParam,
-                               @RequestParam(name = "name", required = false, defaultValue = "world") String nameParam) {
+    public JsonGreeting getGreeting(@RequestParam(name = "salutation", required = false, defaultValue = "hello") String salutationParam,
+                                    @RequestParam(name = "name", required = false, defaultValue = "world") String nameParam) {
 
         log.info("A greeting was requested");
 
-        var greeting = Greeting.of(salutationParam, nameParam);
+        var greeting = JsonGreeting.of(salutationParam, nameParam);
 
         log.info("Greeting returned: {}", greeting);
