+++ src/main/java/com/wilkins/showcase/controller/JsonGreeting.java
@@ -1,0 +1,9 @@
+package com.wilkins.showcase.controller;
+
+import java.io.Serializable;
+
+public record JsonGreeting(String salutation, String name) implements Serializable {
+    public static JsonGreeting of(String salutation, String name) {
+        return new JsonGreeting(salutation, name);
+    }
+}
