--- src/main/java/com/wilkins/showcase/controllers/Greeting.java
@@ -1,9 +1,0 @@
-package com.wilkins.showcase.controllers;
-
-import java.io.Serializable;
-
-public record Greeting(String salutation, String name) implements Serializable {
-    public static Greeting of(String salutation, String name) {
-        return new Greeting(salutation, name);
-    }
-}
